+++ src/ResourceManager/StorageSync/Commands.StorageSync.Test/UnitTests/SystemValidationsProcessorTest.cs
@@ -3,6 +3,7 @@ using Microsoft.Azure.Commands.StorageSync.Evaluation;
 using Microsoft.Azure.Commands.StorageSync.Evaluation.Interfaces;
 using Xunit;
 using Moq;
+using Microsoft.WindowsAzure.Commands.ScenarioTest;
 
 namespace Microsoft.Azure.Commands.StorageSync.Test.UnitTests
 {