+++ src/ResourceManager/StorageSync/Commands.StorageSync/NamespaceValidationsProcessor.cs
@@ -15,16 +15,16 @@
         #region Constructors
         public NamespaceValidationsProcessor(IList<INamespaceValidation> validations, IList<IOutputWriter> outputWriters, IProgressReporter progressReporter)
         {
-            _validations = validations;
-            _outputWriters = outputWriters;
-            _progressReporter = progressReporter;
+            this._validations = validations;
+            this._outputWriters = outputWriters;
+            this._progressReporter = progressReporter;
         }
         #endregion
 
         #region Public methods
         public void BeginDir(IDirectoryInfo node)
         {
-            foreach (INamespaceValidation validation in _validations)
+            foreach (INamespaceValidation validation in this._validations)
             {
                 IValidationResult validationResult = validation.Validate(node);
                 Broadcast(validationResult);
@@ -33,12 +33,12 @@
 
         public void EndDir(IDirectoryInfo node)
         {
-            _progressReporter.CompleteStep();
+            this._progressReporter.CompleteStep();
         }
 
         public void EndOfEnumeration(INamespaceInfo namespaceInfo)
         {
-            foreach (INamespaceValidation validation in _validations)
+            foreach (INamespaceValidation validation in this._validations)
             {
                 IValidationResult validationResult = validation.Validate(namespaceInfo);
                 Broadcast(validationResult);
@@ -47,12 +47,12 @@
 
         public void NextFile(IFileInfo node)
         {
-            foreach (INamespaceValidation validation in _validations)
+            foreach (INamespaceValidation validation in this._validations)
             {
                 IValidationResult validationResult = validation.Validate(node);
                 Broadcast(validationResult);
             }
-            _progressReporter.CompleteStep();
+            this._progressReporter.CompleteStep();
         }
 
         public void UnauthorizedDir(IDirectoryInfo dir)
@@ -62,14 +62,14 @@
 
         public void NamespaceHint(long directoryCount, long fileCount)
         {
-            _progressReporter.ReserveSteps(directoryCount + fileCount);
+            this._progressReporter.ReserveSteps(directoryCount + fileCount);
         }
         #endregion
 
         #region Private methods
         private void Broadcast(IValidationResult validationResult)
         {
-            foreach (IOutputWriter outputWriter in _outputWriters)
+            foreach (IOutputWriter outputWriter in this._outputWriters)
             {
                 outputWriter.Write(validationResult);
             }