+++ src/ResourceManager/StorageSync/Commands.StorageSync/Validations/SystemValidations/OSVersionValidation.cs
@@ -16,42 +16,42 @@
         #region Constructors
         public OSVersionValidation(IConfiguration configuration) : base(configuration, "OS version check", ValidationType.OsVersion)
         {
-            _osversions = new Dictionary<string, string>();
-            _osversions["10.0"] = "Windows Server 2016";
-            _osversions["6.3"] = "Windows Server 2012 R2";
+            this._osversions = new Dictionary<string, string>();
+            this._osversions["10.0"] = "Windows Server 2016";
+            this._osversions["6.3"] = "Windows Server 2012 R2";
 
-            _editions = new Dictionary<uint, string>();
-            _editions[0] = "Undefined";
-            _editions[1] = "Ultimate Edition";
-            _editions[2] = "Home Basic Edition";
-            _editions[3] = "Home Premium Edition";
-            _editions[4] = "Enterprise Edition";
-            _editions[5] = "Home Basic N Edition";
-            _editions[6] = "Business Edition";
-            _editions[7] = "Standard Server Edition";
-            _editions[8] = "Datacenter Server Edition";
-            _editions[9] = "Small Business Server Edition";
-            _editions[10] = "Enterprise Server Edition";
-            _editions[11] = "Starter Edition";
-            _editions[12] = "Datacenter Server Core Edition";
-            _editions[13] = "Standard Server Core Edition";
-            _editions[14] = "Enterprise Server Core Edition";
-            _editions[15] = "Enterprise Server Edition for Itanium-Based Systems";
-            _editions[16] = "Business N Edition";
-            _editions[17] = "Web Server Edition";
-            _editions[18] = "Cluster Server Edition";
-            _editions[19] = "Home Server Edition";
-            _editions[20] = "Storage Express Server Edition";
-            _editions[21] = "Storage Standard Server Edition";
-            _editions[22] = "Storage Workgroup Server Edition";
-            _editions[23] = "Storage Enterprise Server Edition";
-            _editions[24] = "Server For Small Business Edition";
-            _editions[25] = "Small Business Server Premium Edition";
-            _editions[29] = "Web Server, Server Core";
-            _editions[39] = "Datacenter Edition without Hyper-V, Server Core";
-            _editions[40] = "Standard Edition without Hyper-V, Server Core";
-            _editions[41] = "Enterprise Edition without Hyper-V, Server Core";
-            _editions[42] = "Hyper-V Server";
+            this._editions = new Dictionary<uint, string>();
+            this._editions[0] = "Undefined";
+            this._editions[1] = "Ultimate Edition";
+            this._editions[2] = "Home Basic Edition";
+            this._editions[3] = "Home Premium Edition";
+            this._editions[4] = "Enterprise Edition";
+            this._editions[5] = "Home Basic N Edition";
+            this._editions[6] = "Business Edition";
+            this._editions[7] = "Standard Server Edition";
+            this._editions[8] = "Datacenter Server Edition";
+            this._editions[9] = "Small Business Server Edition";
+            this._editions[10] = "Enterprise Server Edition";
+            this._editions[11] = "Starter Edition";
+            this._editions[12] = "Datacenter Server Core Edition";
+            this._editions[13] = "Standard Server Core Edition";
+            this._editions[14] = "Enterprise Server Core Edition";
+            this._editions[15] = "Enterprise Server Edition for Itanium-Based Systems";
+            this._editions[16] = "Business N Edition";
+            this._editions[17] = "Web Server Edition";
+            this._editions[18] = "Cluster Server Edition";
+            this._editions[19] = "Home Server Edition";
+            this._editions[20] = "Storage Express Server Edition";
+            this._editions[21] = "Storage Standard Server Edition";
+            this._editions[22] = "Storage Workgroup Server Edition";
+            this._editions[23] = "Storage Enterprise Server Edition";
+            this._editions[24] = "Server For Small Business Edition";
+            this._editions[25] = "Small Business Server Premium Edition";
+            this._editions[29] = "Web Server, Server Core";
+            this._editions[39] = "Datacenter Edition without Hyper-V, Server Core";
+            this._editions[40] = "Standard Edition without Hyper-V, Server Core";
+            this._editions[41] = "Enterprise Edition without Hyper-V, Server Core";
+            this._editions[42] = "Hyper-V Server";
         }
         #endregion
 
@@ -76,7 +76,7 @@
 
             if (!IsValidVersion(osVersion))
             {
-                string supportedVersions = String.Join(", ", this.Configuration.ValidOsVersions().Where(o => _osversions.ContainsKey(o)).Select(o => _osversions[o]));
+                string supportedVersions = String.Join(", ", this.Configuration.ValidOsVersions().Where(o => this._osversions.ContainsKey(o)).Select(o => this._osversions[o]));
 
                 return new ValidationResult()
                 {
@@ -89,8 +89,8 @@
 
             if (!IsValidSKU(sku))
             {
-                string skuEdition = _editions.ContainsKey(sku) ? _editions[sku] : sku.ToString();
-                string supportedSKU = String.Join(", ", this.Configuration.ValidOsSKU().Where(o => _editions.ContainsKey(o)).Select(o => _editions[o]));
+                string skuEdition = this._editions.ContainsKey(sku) ? this._editions[sku] : sku.ToString();
+                string supportedSKU = String.Join(", ", this.Configuration.ValidOsSKU().Where(o => this._editions.ContainsKey(o)).Select(o => this._editions[o]));
 
                 return new ValidationResult()
                 {